Refactor project layout to use `apps` directory and extract UI logic
- Moved `accounts`, `dadjokes`, `news`, and `notifications` into a new `apps/` directory to clean up the root project.
- Created a new `frontend` application within `apps/` to decouple presentation logic from domain apps.
- Extracted UI views (`newspaper_frontpage`, `joke_of_the_day`, `previous_jokes`) into `frontend.views`.
- Kept API and management endpoints (`rate_joke`, `manage_sources`, `delete_source`) in their native apps.
- Updated `settings.py` (`INSTALLED_APPS`), inter-app imports, and URL routing (`urls.py`) to align with the new `apps.` namespace and structure.
- Adjusted Celery beat task paths and test seeding scripts appropriately.
