+++ apps/accounts/admin.py
