+++ apps/dadjokes/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DadjokesConfig(AppConfig):
-    name = "dadjokes"
+    name = "apps.dadjokes"
