+++ apps/frontend/admin.py
