+++ apps/news/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class NewsConfig(AppConfig):
-    name = "news"
+    name = "apps.news"
