+++ apps/news/models.py
