+++ apps/notifications/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class NotificationsConfig(AppConfig):
-    name = "notifications"
+    name = "apps.notifications"
