+++ apps/notifications/tasks.py
@@ -2,7 +2,7 @@ from celery import shared_task
 from django.utils import timezone
 from django.core.mail import send_mail
 from .models import NotificationPreference, NotificationLog
-from dadjokes.models import Joke
+from apps.dadjokes.models import Joke
 import pytz
 
 @shared_task
