+++ apps/notifications/templates/notifications/manage.html
